Adding the test, data, and backup containers
The test container will run all Mocha tests and output the test data to
/var/lib/koop.

The data container is used as a Data Volume container, which will hold
the test results along with any Koop generated data from the export
worker.

The backup container mounts the data container and creates a backup of
the Koop data directory. After executing the backup container you can
use docker to export the container.

Also did a rework of the config/scripts/koop.sh script.

- Reworked how subcommands are being handled
- Renamed re-create to recreate
- Added the export command

The export command will export and extract the contents of the
/var/lib/koop folder from the data container.